Created DatabaseConnectorTests and subsequently fixed DatabaseConnector
Fixed the way values are read from the select command. Also created test to Assert first row of database.

+++ TradingLib/DatabaseRow.cs
@@ -2,12 +2,12 @@
 {
 	public class DatabaseRow
 	{
-		public int date;
+		public double date;
 		public double price;
 		public double volBid;
 		public double volAsk;
 
-		public DatabaseRow(int date, double price, double volBid, double volAsk)
+		public DatabaseRow(double date, double price, double volBid, double volAsk)
 		{
 			this.date = date;
 			this.price = price;
